Add thread-current-inl.h to solve the Undefined symbols error
